Great leaders are willing to sacrifice the numbers to save the people. Poor leaders sacrifice the people to save the numbers

+++ pointers_arrays_strings/100-atoi.c
@@ -6,7 +6,7 @@
  */
 int _atoi(char *s)
 {
-	long int i, j, p, n, c;
+	int i, j, p, n, c;
 
 	n = 0;
 	p = 1;
